requirements.txt
ib-insync>=0.9.86
google-generativeai
requests
pandas
pytz
matrix-nio
